Time: 0 ms (100.00%), Space: 10.8 MB (54.85%) - LeetHub

+++ 0080-remove-duplicates-from-sorted-array-ii/0080-remove-duplicates-from-sorted-array-ii.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int removeDuplicates(vector<int>& n) {
         int sz=0;
-        int cnt=1;
+        short cnt=1;
         int ptr=0;
         for(int i=1;i<n.size();i++){
             if(cnt==2){
